+++ tcex/app_config_object/__init__.py
@@ -3,3 +3,5 @@
 # flake8: noqa
 from .install_json import InstallJson
 from .layout_json import LayoutJson
+from .profile import Profile, ProfileInteractive
+from .tcex_json import TcexJson