+++ tcex/testing/test_case_playbook.py
@@ -17,12 +17,7 @@ class TestCasePlaybook(TestCasePlaybookCommon):
         Returns:
             [type]: [description]
         """
-        # resolve env vars
-        for k, v in args.items():
-            if isinstance(v, str):
-                args[k] = self.resolve_env_args(v)
-
-        args['tc_playbook_out_variables'] = ','.join(self.output_variables)
+        args['tc_playbook_out_variables'] = self.ij.output_variable_array
         self.log_data('run', 'args', args)
         self.app = self.app_init(args)
 
@@ -75,21 +70,13 @@ class TestCasePlaybook(TestCasePlaybookCommon):
 
         return self._exit(self.app.tcex.exit_code)
 
-    def run_profile(self, profile):
+    def run_profile(self):
         """Run an App using the profile name."""
-        if isinstance(profile, str):
-            profile = self.init_profile(profile)
-
-        # build args from install.json
-        args = {}
-        args.update(profile.get('inputs', {}).get('required', {}))
-        args.update(profile.get('inputs', {}).get('optional', {}))
-
         # run the App
-        exit_code = self.run(args)
+        exit_code = self.run(self.profile.args)
 
         # add context for populating output variables
-        self._context_tracker.append(self.context)
+        self.profile.add_context(self.context)
 
         return exit_code
 
@@ -97,4 +84,5 @@ class TestCasePlaybook(TestCasePlaybookCommon):
         """Run before each test method runs."""
         super().setup_method()
         self.stager.redis.from_dict(self.redis_staging_data)
+
         self.redis_client = self.tcex.redis_client